pyproject.toml
[project]
name = "allscrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "pywhatkit>=5.4",
    "rich>=13.9.4",
    "telethon>=1.37.0",
]
